* Merge in the resync branch, to continue development of it in trunk.
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@3327 37977431-3df6-0310-b722-df95706aa16b

+++ rts/System/creg/STL_List.h
@@ -14,7 +14,7 @@ namespace creg {
 		void Serialize (ISerializer *s, void *inst) {
 			T& ct = *(T*)inst;
 			if (s->IsWriting ()) {
-				int size = ct.size();
+				int size = (int)ct.size();
 				s->Serialize (&size,sizeof(int));
 				for (typename T::iterator it = ct.begin(); it!=ct.end(); ++it)
 					elemType->Serialize (s, &*it);
